+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/main/java/io/flutter/plugins/googlemaps/Convert.java
@@ -599,7 +599,15 @@ class Convert {
     sink.setPosition(toLatLng(marker.getPosition().toList()));
     sink.setRotation(marker.getRotation().floatValue());
     sink.setVisible(marker.getVisible());
-    sink.setZIndex(marker.getZIndex().floatValue());
+
+    // We compare with zero because zIndexInt if non zero should get higher priority than zIndex
+    int zIndexInt = marker.getZIndexInt().intValue();
+    float zIndex = marker.getZIndex().floatValue();
+    if (zIndexInt != 0) {
+      sink.setZIndex(zIndexInt);
+    } else {
+      sink.setZIndex(zIndex);
+    }
   }
 
   private static void interpretInfoWindowOptions(