+++ packages/google_maps_flutter/google_maps_flutter_web/example/integration_test/marker_test.dart
@@ -6,6 +6,7 @@ import 'dart:async';
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:google_maps/google_maps.dart' as gmaps;
+import 'package:google_maps_flutter_platform_interface/google_maps_flutter_platform_interface.dart';
 import 'package:google_maps_flutter_web/google_maps_flutter_web.dart';
 // ignore: implementation_imports
 import 'package:google_maps_flutter_web/src/utils.dart';
@@ -204,6 +205,22 @@ void main() {
           controller.hideInfoWindow();
         }, throwsAssertionError);
       });
+    });
+
+    testWidgets('interpretCorrectZIndexInConversion',
+        (WidgetTester tester) async {
+      final gmaps.Marker newGmMarker = gmaps.Marker();
+      const Marker marker = Marker(
+        markerId: MarkerId('1'),
+        zIndex: 2.0,
+        zIndexInt: 4,
+      );
+
+      final gmaps.MarkerOptions markerOptions =
+          await markerOptionsFromMarker(marker, newGmMarker);
+      final gmaps.Marker gmMarker = gmaps.Marker(markerOptions);
+
+      expect(gmMarker.zIndex, 4);
     });
   });
 }