+++ packages/google_maps_flutter/google_maps_flutter/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 2.9.1
 
+* Deprecates zIndex parameter in Marker and replace with zIndexInt.
 * Updates minimum supported SDK version to Flutter 3.19/Dart 3.3.
 
 ## 2.9.0