+++ packages/google_maps_flutter/google_maps_flutter_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.14.12
+
+* Deprecates zIndex parameter in Marker and replace with zIndexInt.
+
 ## 2.14.11
 
 * Updates internal Pigeon API to use newer features.